--- Frontend/Models/Features/FeaturesCommonViewModel.cs
@@ -1,9 +1,0 @@
-namespace Frontend.Models.Features
-{
-    public class FeaturesCommonViewModel
-    {
-        public string Urn { get; set; }
-        public bool ReturnToPreview { get; set; }
-        public string OutgoingAcademyName { get; set; }
-    }
-}
