+++ Frontend/Pages/Projects/Features/Reason.cshtml
@@ -1,5 +1,4 @@
 ﻿@page "/project/{urn}/features/reason"
-@using Microsoft.AspNetCore.Mvc.ModelBinding
 @model Frontend.Pages.Projects.Features.Reason
 
 @{
