+++ Frontend/Views/Shared/_RadioButtons.cshtml
@@ -1,10 +1,11 @@
+@using Helpers
 @model List<Frontend.Models.Forms.RadioButtonViewModel>
 
 <div class="govuk-radios">
     @for (var i = 0; i < Model.Count(); i++)
     {
         var radioButton = Model[i];
-        var nameOrValue = (i == 0 ? radioButton.Name : radioButton.Value);
+        var nameOrValue = (i == 0 ? radioButton.Name.ToHtmlName() : radioButton.Value);
         <div class="govuk-radios__item">
             <input class="govuk-radios__input" id="@nameOrValue" name="@radioButton.Name" type="radio" value="@radioButton.Value" checked="@radioButton.Checked">
             <label class="govuk-label govuk-radios__label" for="@nameOrValue">
